mongo-seed/requirements.txt
dnspython==2.6.1
mongoengine==0.28.2
pymongo==4.8.0
python-dotenv==1.0.1
